GSvar: moved creation of IGV sessions to LoginManager to assign IGV ports after login correctly
