+++ src/GSvar/GlobalServiceProvider.h
@@ -6,14 +6,6 @@
 #include "DatabaseService.h"
 #include "StatisticsService.h"
 #include <QDialog>
-#include <QMutex>
-
-struct IGVSession
-{
-	int port;
-	bool is_initialized;
-	QString genome;
-};
 
 ///Provider class for GSvar-wide services
 class GlobalServiceProvider
@@ -43,16 +35,6 @@ public:
 	static void executeCommandInIGV(QString command, bool init_if_not_done, int session_index);
 	static void gotoInIGV(QString region, bool init_if_not_done, int session_index = 0);
 	static void loadFileInIGV(QString filename, bool init_if_not_done, bool is_virus_genome = false);
-	//methods to handle dedicated IGV instances
-	static int createIGVSession(int port, bool is_initialized, QString genome);
-	static void removeIGVSession(int session_index);
-	static int findAvailablePortForIGV();
-	static int getIGVPort(int session_index);
-	static void setIGVPort(int port, int session_index);
-	static QString getIGVGenome(int session_index);
-	static void setIGVGenome(QString genome, int session_index);
-	static bool isIGVInitialized(int session_index);
-	static void setIGVInitialized(bool is_initialized, int session_index);
 
 	//opening GSvar files
 	static void openGSvarViaNGSD(QString processed_sample_name, bool search_multi);
@@ -72,8 +54,6 @@ private:
 	QSharedPointer<FileLocationProvider> file_location_provider_;
 	QSharedPointer<DatabaseService> database_service_;
 	QSharedPointer<StatisticsService> statistics_service_;
-	QList<IGVSession> session_list_;
-	QMutex mutex_;
 };
 
 #endif // GLOBALSERVICEPROVIDER_H
